Admin - move action ChangeStore from SettingsController to the HomepageController
